try catch and toast notification added

+++ front_end/src/helpers/adminAxios.js
@@ -22,16 +22,19 @@ adminAxiosIntance.interceptors.request.use((req) => {
 })
 
 
+
 adminAxiosIntance.interceptors.response.use((res) => {
   return res;
 }, (error) => {
-  console.log(error);
-  const { status } = error.response;
-  if(status === 500) {
-    localStorage.clear();
-    store.dispatch({ type: authConstants.LOGOUT_SUCCESS});
+  try{
+    const { status } = error.response ? error.response.status : 500;
+    if(status && status === 500) {
+      localStorage.clear();
+      store.dispatch({ type: authConstants.ADMIN_LOGOUT_SUCCESS});
+    }
+  }catch(error){
+    return Promise.reject(error);
   }
-  return Promise.reject(error);
 })
 
 export default adminAxiosIntance;
